refactor: rename components/AnswersHistory, useAnsweredHistory hook
