+++ src/pages/AnswersHistoryPage.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import Tabs from "../components/common/Tabs";
 import { useState } from "react";
-import Answers from "../components/AnswersHistory/Answers";
+import Answers from "../components/AnsweredHistory/Answers";
 import { useCategory } from "../hooks/UseCategory";
 import { useAnsweredHistory } from "../hooks/UseAnsweredHistory";
